+++ src/main/java/gov/epa/bencloud/server/database/jooq/data/tables/AirQualityCell.java
@@ -65,7 +65,7 @@ public class AirQualityCell extends TableImpl<AirQualityCellRecord> {
     /**
      * The column <code>data.air_quality_cell.grid_cell_id</code>.
      */
-    public final TableField<AirQualityCellRecord, Long> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.BIGINT, this, "");
+    public final TableField<AirQualityCellRecord, Integer> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.INTEGER, this, "");
 
     /**
      * The column <code>data.air_quality_cell.metric_id</code>.
@@ -186,7 +186,7 @@ public class AirQualityCell extends TableImpl<AirQualityCellRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row8<Integer, Integer, Integer, Long, Integer, Integer, String, BigDecimal> fieldsRow() {
+    public Row8<Integer, Integer, Integer, Integer, Integer, Integer, String, BigDecimal> fieldsRow() {
         return (Row8) super.fieldsRow();
     }
 }
