+++ src/main/java/gov/epa/bencloud/server/database/jooq/data/tables/CrosswalkEntry.java
@@ -66,7 +66,7 @@ public class CrosswalkEntry extends TableImpl<CrosswalkEntryRecord> {
     /**
      * The column <code>data.crosswalk_entry.source_grid_cell_id</code>.
      */
-    public final TableField<CrosswalkEntryRecord, Long> SOURCE_GRID_CELL_ID = createField(DSL.name("source_grid_cell_id"), SQLDataType.BIGINT, this, "");
+    public final TableField<CrosswalkEntryRecord, Integer> SOURCE_GRID_CELL_ID = createField(DSL.name("source_grid_cell_id"), SQLDataType.INTEGER, this, "");
 
     /**
      * The column <code>data.crosswalk_entry.target_col</code>.
@@ -81,7 +81,7 @@ public class CrosswalkEntry extends TableImpl<CrosswalkEntryRecord> {
     /**
      * The column <code>data.crosswalk_entry.target_grid_cell_id</code>.
      */
-    public final TableField<CrosswalkEntryRecord, Long> TARGET_GRID_CELL_ID = createField(DSL.name("target_grid_cell_id"), SQLDataType.BIGINT, this, "");
+    public final TableField<CrosswalkEntryRecord, Integer> TARGET_GRID_CELL_ID = createField(DSL.name("target_grid_cell_id"), SQLDataType.INTEGER, this, "");
 
     /**
      * The column <code>data.crosswalk_entry.percentage</code>.
@@ -162,7 +162,7 @@ public class CrosswalkEntry extends TableImpl<CrosswalkEntryRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row8<Integer, Integer, Integer, Long, Integer, Integer, Long, BigDecimal> fieldsRow() {
+    public Row8<Integer, Integer, Integer, Integer, Integer, Integer, Integer, BigDecimal> fieldsRow() {
         return (Row8) super.fieldsRow();
     }
 }
