+++ src/main/java/gov/epa/bencloud/server/database/jooq/data/tables/GetVariable.java
@@ -50,7 +50,7 @@ public class GetVariable extends TableImpl<GetVariableRecord> {
     /**
      * The column <code>data.get_variable.grid_cell_id</code>.
      */
-    public final TableField<GetVariableRecord, Long> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.BIGINT, this, "");
+    public final TableField<GetVariableRecord, Integer> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.INTEGER, this, "");
 
     /**
      * The column <code>data.get_variable.value</code>.
@@ -122,7 +122,7 @@ public class GetVariable extends TableImpl<GetVariableRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row3<String, Long, BigDecimal> fieldsRow() {
+    public Row3<String, Integer, BigDecimal> fieldsRow() {
         return (Row3) super.fieldsRow();
     }
